requirements.txt
# YouTube API
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# OpenAI
openai==1.3.0

# Web scraping & data
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.11
youtube-transcript-api==0.6.1

# Trends & social monitoring
pytrends==4.9.2
praw==7.7.1

# Media generation
edge-tts==6.1.9
pillow==10.1.0

# RAG (leverage existing youtube-automation-rag)
chromadb==0.4.22
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
